Rename db connect file
