+++ src/app/account/manage-suppliers/page.tsx
@@ -4,7 +4,7 @@ import Link from 'next/link'
 import { redirect } from 'next/navigation'
 
 import { Button } from '@/components/ui/button'
-import { db } from '@/db/db'
+import { db } from '@/db/connection'
 import * as schema from '@/db/schema'
 import { getAuthUserId } from '@/utils/auth'
 import { valueToPretty } from '@/utils/enum-helpers'
