+++ src/models/user.ts
@@ -1,6 +1,6 @@
 import { eq } from 'drizzle-orm'
 
-import { db } from '@/db/db'
+import { db } from '@/db/connection'
 import * as schema from '@/db/schema'
 import { InsertUserDetailRaw, SetUserDetailRaw, UserDetailRaw } from '@/models/types'
 import { emptyStringToNullIfAllowed } from '@/utils/empty-strings'
